OK, Chris & I hacked on this and got it to build on Windows. We ran the tests and all succeeded except for CPodesIntegratorTest, in which cpodes returned an apparently unexpected status of -4. But we'll let Peter E sort this out ...

+++ trunk/Integrator/src/CPodesIntegratorRep.cpp
@@ -29,8 +29,8 @@
  * USE OR OTHER DEALINGS IN THE SOFTWARE.                                     *
  * -------------------------------------------------------------------------- */
 
-#include "simmath/internal/CPodesIntegratorRep.h"
-//#include "SimTKcpodes/cpodes/cpodes.h"
+#include "SimTKcommon.h"
+#include "CPodesIntegratorRep.h"
 
 using namespace SimTK;
 
@@ -337,7 +337,7 @@ Integrator::SuccessfulStepStatus CPodesIntegratorRep::stepTo(Real reportTime, Re
             findEventIds(eventIds);
             std::vector<Real> eventTimes(eventIds.size());
             std::vector<EventStatus::EventTrigger> eventTransitions(eventIds.size());
-            for (int i = 0; i < eventIds.size(); ++i) {
+            for (int i = 0; i < (int)eventIds.size(); ++i) {
                 eventTimes[i] = tret;
                 eventTransitions[i] = EventStatus::AnySignChange;
             }
